Move core::bool tests to run-pass
These are closer to language tests than library.
